remove server partials, adjust gulp for preprocessing index.html
