+++ lib/tasks/inject.js
@@ -2,6 +2,17 @@ var bower = require('main-bower-files');
 
 module.exports = function($, gulp, paths){
 
+  var envConfigDevelopment = {
+    BASE_HREF: 'localhost:3000'
+  };
+
+  var envConfigProduction = {
+    BASE_HREF: 'mockr-hr.herokuapp.com'
+  };
+
+  var envConfig = (process.env.NODE_ENV === 'production') ? envConfigProduction
+                                                          : envConfigDevelopment;
+
   return {
     /**
      * @name            dev
@@ -15,6 +26,7 @@ module.exports = function($, gulp, paths){
      *                  Inject all .css files into the index.html between the <!-- app:css --> comment
      *                  Inject all vendor files into the index.html between the <!-- vendor:css && vendor:js --> comments
      *                  Inject all template files into the index.html between the <!-- templates:js --> comment
+     *                  Preprocess base href into the index.html
      *                  Pipe the index.html into the ./build/ directory
      * @return {Stream}
      */
@@ -32,6 +44,7 @@ module.exports = function($, gulp, paths){
         .pipe($.inject(css      , {name:'app'   , ignorePath:'build'}))
         .pipe($.inject(vendor   , {name:'vendor', ignorePath:'build'}))
         .pipe($.inject(templates, {name:'templates'}))
+        .pipe($.preprocess({ context: envConfig }))
         .pipe( dest )
     },
     /**
@@ -63,6 +76,7 @@ module.exports = function($, gulp, paths){
         .pipe($.inject(css      , {name:'app'   , ignorePath:'build'}))
         .pipe($.inject(vendor   , {name:'vendor', ignorePath:'build'}))
         .pipe($.inject(templates, {name:'templates'}))
+        .pipe($.preprocess({ context: envConfig }))
         .pipe( dest )
     }
   };
