+++ src/client/index.html
@@ -1,7 +1,22 @@
 <!DOCTYPE html>
 <html data-ng-app="app">
 <head>
-  <title></title>
+  <meta charset="utf-8">
+  <meta http-equiv="X-UA-Compatible" content="IE=edge">
+  <title>Mockr</title>
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <base href="http://<!-- @echo BASE_HREF -->/">
+  <meta property="og:site_name" content="Mockr">
+  <meta name="twitter:card" content="summary">
+  <meta name="description" content="">
+  <meta property="og:type" content="website">
+  <meta property="og:url" content="">
+  <meta property="og:title" content="">
+  <meta property="og:image" content="">
+  <meta name="twitter:url" content="">
+  <meta name="twitter:title" content="">
+  <meta name="twitter:description" content="">
+  <meta name="twitter:image" content="">
 <!-- vendor:css -->
 <!-- endinject -->
