+++ src/server/routes.js
@@ -1,13 +1,15 @@
 var router = require('express').Router();
 
 module.exports = function applicationRouter(app) {
-	
+
+  // Mount routes to router
   require('./api/user')(router);
   require('./api/endpoint')(router);
 
-  // router.get('/*', function(req, res, next) {
-  //   res.render('index');
-  // });
+  // Catch all routes and send to Angular for handling
+  router.get('/*', function(req, res, next) {
+    res.redirect('/');
+  });
 
   app.use(router);
 };
