--- src/server/views/_partials/footer.jade
@@ -1,1 +1,0 @@
-footer(role="footer" id="footer")
