--- src/server/views/_partials/head.template.jade
@@ -1,34 +1,0 @@
-meta(charset="utf-8")
-meta(http-equiv="X-UA-Compatible" content="IE=edge")
-title Archivr
-meta(name="viewport" content="width=device-width, initial-scale=1")
-
-//- base URL for angular
-base(href="http:///* @echo BASE_HREF *//")
-
-//- these meta properties do not change
-meta(property="og:site_name" content="Archivr")
-meta(name="twitter:card" content="summary")
-
-//- meta block is unqiue on a page by page basis.
-block meta
-  //- meta description for each page
-  meta(name="description" content="")
-
-  //- og meta is for facebook
-  meta(property="og:type" content="website")
-  meta(property="og:url" content="")
-  meta(property="og:title" content="")
-  meta(property="og:image" content="")
-
-  //- twitter meta
-  meta(name="twitter:url" content="")
-  meta(name="twitter:title" content="")
-  meta(name="twitter:description" content="")
-  meta(name="twitter:image" content="")
-
-link(rel="stylesheet" href="//maxcdn.bootstrapcdn.com/bootstrap/3.3.2/css/bootstrap.min.css")
-link(rel="stylesheet" href="css/app.css")
-
-block header_scripts
-  script(src="/js/app.min.js")
