--- src/server/views/index.jade
@@ -1,15 +1,0 @@
-doctype html
-html(lang="en" class="no-js" ng-app="Archivr")
-
-  
-
-  
-
-    main(role="main" id="main" class="container")
-
-      block content
-        div(ui-view)
-
-   
-
-  block footer_scripts
